requirements.txt
fastapi
uvicorn
pydantic
psycopg2-binary
pandas
openpyxl
requests
meta-ads-collector
google-generativeai
langdetect
pycountry
python-dotenv